local storage
In this commit, local storage was added to the project so that user todos and projects are saved after leaving/refreshing the page.

I also decided to change the mehtods on the classes from arrow functions to regular function declarations after learning that using arrow fns will duplicate them every time the class is instantiated.